Forbid unauthenticated requests and refactor request routing
