+++ src/main/java/com/filipovski/server/utils/Utils.java
@@ -4,6 +4,7 @@ import io.netty.util.AttributeKey;
 
 public class Utils {
     public static final String proxySessionName = "poidSESSION";
+
     public static final AttributeKey sessionAttributeKey  = AttributeKey.valueOf("session");
 
     public static boolean notForLocalServer(String host) {
@@ -13,6 +14,4 @@ public class Utils {
     public static boolean isAuthenticated(String name) {
         return true;
     }
-
-
 }
